requirements.txt
Flask==2.3.2
Flask-WTF==1.1.1
Jinja2==3.1.2
eventlet==0.33.3
mysqlclient==2.2.0
Pillow==8.3.2
